requirements.txt
python-chess
numpy
tqdm
pygame
